requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==2.3.7
itsdangerous==2.1.2 